Refactor snippets. WIP
