+++ great_expectations/render/base.py
@@ -1,23 +1,22 @@
 class Renderer(object):
-    def __init__(self, *args, **kwargs):
-        """
-        Note: This shouldn't get used much other than input validation. We expect most renderers to be stateless, single-use classes.
-        """
-        self.validate_input(*args, **kwargs)
 
-    def validate_input(self, *args, **kwargs):
+    @classmethod
+    def validate_input(cls, *args, **kwargs):
         raise NotImplementedError
 
-    def render(self):
-        pass
-
-class SingleExpectationRenderer(Renderer):
-    def __init__(self, expectation):
-        self.expectation = expectation
-
-    def validate_input(self, expectation):
-        return True
-
-    def render(self):
+    @classmethod
+    def _render_to_json(cls, *args, **kwargs):
         raise NotImplementedError
 
+    @classmethod
+    def render(cls, *args, **kwargs):
+        raise NotImplementedError
+
+
+class ViewModelRenderer(Renderer):
+
+    @classmethod
+    def render(cls, *args, **kwargs):
+        class.validate_input(*args, **kwargs)
+        class._render_to_json(*args, **kwargs)
+        raise NotImplementedError
