--- great_expectations/render/view_models/default_html/fixtures/templates/bk/base.j2
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Data documentation compiled by Great Expectations</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>{% block title %}{% endblock %}</title>
-
-    {# {# Remove this when not debugging: #}
-    <meta http-equiv="refresh" content="1"/>
-
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"/>
-    <style>
-      body {
-        position: relative;
-      }
-      .container {
-        margin-top: 50px;
-      }
-      .sticky {
-        position: -webkit-sticky;
-        position: sticky;
-        max-height: 90vh;
-        overflow-y: auto;
-        top: 15px;
-      }
-      .param-span{
-        background:#eef;
-        padding:5px;
-      }
-    </style>
-  </head>
-
-  <body
-        data-spy="scroll"
-        data-target="#navigation"
-        data-offset="50"
-    >
-
-    {% include 'ge_info.j2' %}
-
-    <div class="container">
-      <div class="row">
-
-        <div class="col-3">
-          {% block navbar %}{% endblock %}
-        </div>
-
-        <div class="col-9">
-          {% block sections %}{% endblock %}
-        </div>
-
-      </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-  </body>
-</html>
