--- great_expectations/render/view_models/default_html/fixtures/templates/bk/mock.j2
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Data documentation compiled by Great Expectations</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>{% block title %}{% endblock %}</title>
-    <meta http-equiv="refresh" content="1"/>
-    <link
-      rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-    />
-    <style>
-      body {
-        position: relative;
-      }
-      .container {
-        margin-top: 50px;
-      }
-      .sticky {
-        position: -webkit-sticky;
-        position: sticky;
-        max-height: 90vh;
-        overflow-y: auto;
-        top: 15px;
-      }
-    </style>
-  </head>
-
-  <body
-        data-spy="scroll"
-        data-target="#navigation"
-        data-offset="50"
-    >
-    <div style="float: right; width:240px; margin: 20px;">
-      <img
-        src="https://great-expectations.readthedocs.io/en/latest/_images/generic_dickens_protagonist.png"
-      />
-      Documentation autogenerated using&nbsp;
-      <a href="https://github.com/great-expectations/great_expectations">
-        Great Expectations</a
-      >. This is a beta feature. YMMV.
-    </div>
-    <div class="container">
-      <div class="row">
-        <div class="col-3">
-          <nav id="navigation" class="navbar navbar-light bg-light sticky">
-            <nav class="nav nav-pills flex-column">
-              <a class="navbar-brand" href="#">MyTable</a>
-              {% for i in range(20) %}
-                <a class="nav-link" href="#section-{{i}}">Section {{i}}</a>
-              {% endfor %}
-            </nav>
-          </nav>
-        </div>
-        <div class="col-9">
-            {% for i in range(20) %}
-                <div id="section-{{i}}">
-                    <h4>Section {{ i }}</h4>
-                    <p>{{ lipsum(1) }}</p>
-                    <img style="max-width:50%; max-height:200px; float: right;" src="http://localhost:8000/plot-{{range(0,10)|random}}.png"/>
-                    <p>
-                        <ul>
-                            {% for i in range(7) %}
-                            <li>x_var is a required field.</li>
-                            {% endfor %}
-                            <li>y_var is a required field.</li>
-                            <li>x_var must never be missing.</li>
-                            <li>y_var must never be missing.</li>
-                        </ul>
-                    </p>
-                    <p>{{ lipsum(1) }}</p>
-                </div>
-            {% endfor %}
-          </div>
-        </div>
-      </div>
-    </div>
-    <script
-      src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
-      integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
-      integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
-      integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
-      crossorigin="anonymous"
-    ></script>
-    <script src="src/index.js" />
-  </body>
-</html>
