--- great_expectations/render/view_models/default_html/fixtures/templates/bk/single_page.j2
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Flask Template Example</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="http://netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css" rel="stylesheet" media="screen">
-    <style type="text/css">
-      .container {
-        max-width: 500px;
-        padding-top: 100px;
-      }
-    </style>
-  </head>
-  <body>
-    <div class="container">
-        <ul>
-            {% for item in elements %}
-                <li>{{ item }}</li>
-            {% endfor %}
-        </ul>
-    </div>
-    <script src="http://code.jquery.com/jquery-1.10.2.min.js"></script>
-    <script src="http://netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
-  </body>
-</html>
