--- great_expectations/render/view_models/default_html/fixtures/templates/ge_info.j2
@@ -1,9 +1,0 @@
-<div style="float: right; width:240px; margin: 20px;">
-    <img
-    src="https://great-expectations.readthedocs.io/en/latest/_images/generic_dickens_protagonist.png"
-    />
-    Documentation autogenerated using&nbsp;
-    <a href="https://github.com/great-expectations/great_expectations">
-    Great Expectations</a
-    >. This is a beta feature. YMMV.
-</div>
