--- great_expectations/render/view_models/default_html/fixtures/templates/page.j2
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Data documentation compiled by Great Expectations</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>{% block title %}{% endblock %}</title>
-
-    {# {# Remove this when not debugging: #}
-    <meta http-equiv="refresh" content="2"/>
-
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"/>
-    <style>
-      body {
-        position: relative;
-      }
-      .container {
-        margin-top: 50px;
-      }
-      .sticky {
-        position: -webkit-sticky;
-        position: sticky;
-        max-height: 90vh;
-        overflow-y: auto;
-        top: 15px;
-      }
-      .param-span{
-        background:#eef;
-        padding:5px;
-      }
-
-      .ge-section{
-        display: flex;
-        flex-flow: row wrap;
-        margin-left: -8px; /* Adjustment for the gutter */
-        width: 100%;
-      }
-
-      .ge-section div{
-        flex: auto;
-        {# height: 250px; #}
-        min-width: 150px;
-        margin: 0 8px 8px 0; /* Some gutter */
-        {# background-color: #004;         #}
-        {# border: 1px solid black; #}
-      }
-
-      tr{
-        padding: 2px;
-      }
-      tr.even{
-        background:#ddd;
-      }
-      td{
-        padding: 1px 10px 1px 10px;
-      }
-    </style>
-
-    <script src="https://cdn.jsdelivr.net/npm/vega@5.3.5/build/vega.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vega-lite@3.2.1/build/vega-lite.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vega-embed@4.0.0/build/vega-embed.js"></script>
-  </head>
-
-  <body
-        data-spy="scroll"
-        data-target="#navigation"
-        data-offset="50"
-    >
-
-    {% include 'ge_info.j2' %}
-
-    <div class="container">
-      <div class="row">
-
-        <div class="col-3">
-          {% block navbar %}{% include 'navbar.j2' %}{% endblock %}
-        </div>
-
-        <div class="col-9">
-          {% block sections %}{% include 'sections.j2' %}{% endblock %}
-        </div>
-
-      </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-  </body>
-</html>
