--- great_expectations/render/view_models/default_html/fixtures/templates/section.j2
@@ -1,48 +1,0 @@
-{# !!! This content duplicates sections.j2, with one very minor difference:
-!!! this one is missing `{% for section in sections %}` and `id="section-{{loop.index}}"`
-!!! We'll need to refactor to consolidate. #}
-
-<div style="clear:both; margin-bottom: 50px;">
-    
-    {% for content_block in section["content_blocks"] %}
-        {% if content_block["content_block_type"] == "header" %}
-        <h4>{{content_block["content"][0]}}</h4>
-
-        {% elif content_block["content_block_type"] == "text" %}
-        <p>
-            {{content_block["content"][0]}}
-            {# {{ lipsum(1, min=5, max=20) }} #}
-        </p>
-
-        {% elif content_block["content_block_type"] == "bullet_list" %}
-        <p>
-            <ul>
-                {% for bullet_point in content_block["content"] %}
-                <li>{{ bullet_point }}</li>
-                {% endfor %}
-            </ul>
-        </p>
-
-        {% elif content_block["content_block_type"] == "graph" %}
-        <img
-            style="max-width:32%; max-height:200px; float: right;"
-            src="http://localhost:8000/plot-{{range(5,10)|random}}.png"
-        />
-
-        {% elif content_block["content_block_type"] == "table" %}
-        <table style="width:50%">
-            {% set table = content_block["content"] %}
-            {% for row in table %}
-            <tr class="{{ loop.cycle('odd', 'even') }}" >
-                {% set rowloop = loop %}
-                {% for cell in row %}
-                    <td id="cell-{{ rowloop.index }}-{{ loop.index }}">{{ cell }}</td>
-                {% endfor %}
-            </tr>
-            {% endfor %}
-        </table>
-
-        {% elif content_block["content_block_type"] == "example_list" %}
-        {% endif %}
-    {% endfor %}
-</div>
